requirements.txt
requests
beautifulsoup4
python-telegram-bot==21.4
python-dotenv
